Added testing to ensure that the debut album is actually added to the database in the way we expect.
The tests check that the album is present in the album table and that
every song is present in the database.
Next steps could be to have a test that involves converting each song
into an actual song object in order to ensure that all the provided
information is present.